refactor: Refactor ComponentProductResponse to ComponentResponse
