+++ Server/DelTSZ/Models/Components/ComponentResponse.cs
@@ -1,11 +1,12 @@
 ﻿using DelTSZ.Models.Enums;
+using DelTSZ.Models.Products;
 
-namespace DelTSZ.Models.Products.ComponentProducts;
+namespace DelTSZ.Models.Components;
 
-public class ComponentProductResponse : IProductResponse
+public class ComponentResponse
 {
     public int Id { get; init; }
-    public ComponentProductType ProductType { get; init; }
+    public ComponentType Type { get; init; }
     public DateTime Received { get; init; }
     public double Amount { get; init; }
     public string? UserId { get; init; }
