--- Server/DelTSZ/Models/Products/IProductResponse.cs
@@ -1,9 +1,0 @@
-﻿using DelTSZ.Models.Enums;
-
-namespace DelTSZ.Models.Products;
-
-public interface IProductResponse
-{
-    public int Id { get; init; }
-    public string? UserId { get; init; }
-}
